requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
joblib==1.1.0
nltk==3.7
numpy==1.23.2
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.8.17
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2022.2
